requirements.txt
more_itertools
numpy
torch==1.5.0
tqdm
retry
pytest
pyyaml
tensorboardX
tensorboard
git+https://github.com/Hiroshiba/wandb-client@6c9fa2d48dc48604b37ed330f8c572c2574a3da4
git+https://github.com/Hiroshiba/pytorch-trainer@1.2.0
git+https://github.com/Hiroshiba/acoustic_feature_extractor@7beb1ceb7244a997865778c3a7c730e5fc0cbe02
git+https://github.com/lessw2020/Ranger-Deep-Learning-Optimizer@a170e3dac6443763961d0cb7a16294e4f45f93e4
